More cleanup. Remove obsolete engine procedures.
git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@2720 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/engine/util.c
@@ -458,7 +458,7 @@ const char *
 gnc_locale_default_currency()
 {
   static char currency[4];
-  static gboolean got_it = FALSE;    /* pretty sure "static" is needed */
+  static gboolean got_it = FALSE;
   struct lconv *lc;
   int i;
 
@@ -490,7 +490,7 @@ gnc_locale_decimal_places( void )
     return( places );
 
   lc = gnc_localeconv();
-  places = (int)lc->frac_digits;
+  places = lc->frac_digits;
 
   /* frac_digits is already initialized by gnc_localeconv,
    * hopefully to a reasonable default.                    */
